Atualização na classe de pedidos e tonner, retirada da classe Estoque, remodelagem do gerador de dados de tonners e Inicio da movimentação de Pedidos

+++ Models/Pedido.cs
@@ -25,6 +25,6 @@ namespace TrabalhoWEBMVC3.Models
         public int quantidade { get; set; }
 
         [Display(Name = "Valor")]
-        public float valor { get; set; }
+        public double valor { get; set; }
     }
 }
